+++ src/main/java/net/geforcemods/securitycraft/api/LinkedAction.java
@@ -23,5 +23,10 @@ public enum LinkedAction {
 	/**
 	 * Used when a {@link ModuleType} is removed from an IModuleInventory
 	 */
-	MODULE_REMOVED;
+	MODULE_REMOVED,
+
+	/**
+	 * Used when the {@link Owner} of a block entity changes;
+	 */
+	OWNER_CHANGED;
 }